- Vehicle category manager.
