+++ package.json
@@ -8,10 +8,10 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.38",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
-    "antd": "^4.2.0",
+    "@types/node": "^12.12.39",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
+    "antd": "^4.2.2",
     "axios": "^0.19.2",
     "axios-progress-bar": "^1.2.0",
     "babel-plugin-import": "^1.13.0",
@@ -24,11 +24,11 @@
     "react-app-rewired": "^2.1.6",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3",
-    "typescript": "^3.7.5"
+    "typescript": "^3.9.2"
   },
   "scripts": {
     "start": "react-app-rewired start",
